+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next release
 
  - Add junk detection for AndOrReplacement mutator #509
+ - Add junk detection for MathDiv mutator #511
+ - Add junk detection for MathMul mutator #511
 
 ## [0.2.0] - 02 May 2019