+++ CHANGELOG.md
@@ -8,6 +8,7 @@
    - MathMul #511
    - ReplaceCall #512
  - Handle edge case for AND-OR mutator #501
+ - Add better handling of edge cases when there is no debug information #519, #520, #524
 
 ## [0.2.0] - 02 May 2019